requirements.txt
gpsd-py3==0.3.0
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
scapy==2.5.0
six==1.16.0
tzdata==2024.1
